+++ crates/bevy_gltf/Cargo.toml
@@ -14,6 +14,7 @@ bevy_app = { path = "../bevy_app", version = "0.6.0" }
 bevy_asset = { path = "../bevy_asset", version = "0.6.0" }
 bevy_core = { path = "../bevy_core", version = "0.6.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.6.0" }
+bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.6.0" }
 bevy_pbr = { path = "../bevy_pbr", version = "0.6.0" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.6.0", features = ["bevy"] }
 bevy_render = { path = "../bevy_render", version = "0.6.0" }
