+++ crates/bevy_hierarchy/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "bevy_transform"
+name = "bevy_hierarchy"
 version = "0.6.0"
 edition = "2021"
-description = "Provides transform functionality for Bevy Engine"
+description = "Provides hierarchy functionality for Bevy Engine"
 homepage = "https://bevyengine.org"
 repository = "https://github.com/bevyengine/bevy"
 license = "MIT OR Apache-2.0"
@@ -12,6 +12,8 @@ keywords = ["bevy"]
 # bevy
 bevy_app = { path = "../bevy_app", version = "0.6.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.6.0", features = ["bevy_reflect"] }
-bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.6.0"}
-bevy_math = { path = "../bevy_math", version = "0.6.0" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.6.0", features = ["bevy"] }
+bevy_utils = { path = "../bevy_utils", version = "0.6.0" }
+
+# other
+smallvec = { version = "1.6", features = ["serde", "union", "const_generics"] }
