+++ crates/bevy_hierarchy/src/components/mod.rs
@@ -1,0 +1,5 @@
+mod children;
+mod parent;
+
+pub use children::Children;
+pub use parent::{Parent, PreviousParent};
