+++ crates/bevy_internal/Cargo.toml
@@ -56,6 +56,7 @@ bevy_core = { path = "../bevy_core", version = "0.6.0" }
 bevy_derive = { path = "../bevy_derive", version = "0.6.0" }
 bevy_diagnostic = { path = "../bevy_diagnostic", version = "0.6.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.6.0" }
+bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.6.0" }
 bevy_input = { path = "../bevy_input", version = "0.6.0" }
 bevy_log = { path = "../bevy_log", version = "0.6.0" }
 bevy_math = { path = "../bevy_math", version = "0.6.0" }
