+++ crates/bevy_scene/Cargo.toml
@@ -14,7 +14,7 @@ bevy_app = { path = "../bevy_app", version = "0.6.0" }
 bevy_asset = { path = "../bevy_asset", version = "0.6.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.6.0" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.6.0", features = ["bevy"] }
-bevy_transform = { path = "../bevy_transform", version = "0.6.0" }
+bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.6.0" }
 bevy_utils = { path = "../bevy_utils", version = "0.6.0" }
 
 # other
