+++ crates/bevy_scene/src/command.rs
@@ -4,7 +4,7 @@ use bevy_ecs::{
     system::{Command, Commands},
     world::World,
 };
-use bevy_transform::hierarchy::ChildBuilder;
+use bevy_hierarchy::ChildBuilder;
 
 use crate::{Scene, SceneSpawner};
