--- crates/bevy_transform/.gitignore
@@ -1,24 +1,0 @@
-book/book
-target
-Cargo.lock
-*.log
-
-# Backup files
-.DS_Store
-thumbs.db
-*~
-*.rs.bk
-*.swp
-
-# IDE / Editor files
-*.iml
-.idea
-.vscode
-
-
-#Added by cargo
-#
-#already existing elements are commented out
-
-/target
-**/*.rs.bk
