+++ crates/bevy_transform/src/components/mod.rs
@@ -1,9 +1,5 @@
-mod children;
 mod global_transform;
-mod parent;
 mod transform;
 
-pub use children::Children;
 pub use global_transform::*;
-pub use parent::{Parent, PreviousParent};
 pub use transform::*;
