+++ crates/bevy_transform/src/lib.rs
@@ -3,23 +3,19 @@
 
 /// The basic components of the transform crate
 pub mod components;
-/// Establishing and updating the transform hierarchy
-pub mod hierarchy;
-/// Propagating transform changes down the transform hierarchy
-pub mod transform_propagate_system;
+mod systems;
+pub use crate::systems::transform_propagate_system;
 
 #[doc(hidden)]
 pub mod prelude {
     #[doc(hidden)]
-    pub use crate::{components::*, hierarchy::*, TransformBundle, TransformPlugin};
+    pub use crate::{components::*, TransformBundle, TransformPlugin};
 }
 
 use bevy_app::prelude::*;
-use bevy_ecs::{
-    bundle::Bundle,
-    schedule::{ParallelSystemDescriptorCoercion, SystemLabel},
-};
-use prelude::{parent_update_system, Children, GlobalTransform, Parent, PreviousParent, Transform};
+use bevy_ecs::prelude::*;
+use bevy_hierarchy::HierarchySystem;
+use prelude::{GlobalTransform, Transform};
 
 /// A [`Bundle`] of the [`Transform`] and [`GlobalTransform`]
 /// [`Component`](bevy_ecs::component::Component)s, which describe the position of an entity.
@@ -32,12 +28,12 @@ use prelude::{parent_update_system, Children, GlobalTransform, Parent, PreviousP
 /// ## [`Transform`] and [`GlobalTransform`]
 ///
 /// [`Transform`] is the position of an entity relative to its parent position, or the reference
-/// frame if it doesn't have a [`Parent`](Parent).
+/// frame if it doesn't have a parent.
 ///
 /// [`GlobalTransform`] is the position of an entity relative to the reference frame.
 ///
 /// [`GlobalTransform`] is updated from [`Transform`] in the system
-/// [`transform_propagate_system`](crate::transform_propagate_system::transform_propagate_system).
+/// [`transform_propagate_system`].
 ///
 /// This system runs in stage [`CoreStage::PostUpdate`](crate::CoreStage::PostUpdate). If you
 /// update the[`Transform`] of an entity in this stage or after, you will notice a 1 frame lag
@@ -81,46 +77,33 @@ impl From<Transform> for TransformBundle {
         Self::from_transform(transform)
     }
 }
+/// Label enum for the systems relating to transform propagation
+#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemLabel)]
+pub enum TransformSystem {
+    /// Propagates changes in transform to childrens' [`GlobalTransform`](crate::components::GlobalTransform)
+    TransformPropagate,
+}
+
 /// The base plugin for handling [`Transform`] components
 #[derive(Default)]
 pub struct TransformPlugin;
 
-/// Label enum for the types of systems relating to transform
-#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemLabel)]
-pub enum TransformSystem {
-    /// Propagates changes in transform to childrens' [`GlobalTransform`]
-    TransformPropagate,
-    /// Updates [`Parent`] when changes in the hierarchy occur
-    ParentUpdate,
-}
-
 impl Plugin for TransformPlugin {
     fn build(&self, app: &mut App) {
-        app.register_type::<Children>()
-            .register_type::<Parent>()
-            .register_type::<PreviousParent>()
-            .register_type::<Transform>()
+        app.register_type::<Transform>()
             .register_type::<GlobalTransform>()
-            // add transform systems to startup so the first update is "correct"
+            // Adding these to startup ensures the first update is "correct"
             .add_startup_system_to_stage(
                 StartupStage::PostStartup,
-                parent_update_system.label(TransformSystem::ParentUpdate),
-            )
-            .add_startup_system_to_stage(
-                StartupStage::PostStartup,
-                transform_propagate_system::transform_propagate_system
+                systems::transform_propagate_system
                     .label(TransformSystem::TransformPropagate)
-                    .after(TransformSystem::ParentUpdate),
+                    .after(HierarchySystem::ParentUpdate),
             )
             .add_system_to_stage(
                 CoreStage::PostUpdate,
-                parent_update_system.label(TransformSystem::ParentUpdate),
-            )
-            .add_system_to_stage(
-                CoreStage::PostUpdate,
-                transform_propagate_system::transform_propagate_system
+                systems::transform_propagate_system
                     .label(TransformSystem::TransformPropagate)
-                    .after(TransformSystem::ParentUpdate),
+                    .after(HierarchySystem::ParentUpdate),
             );
     }
 }
